Updated prepare script.

+++ prepare.sh
@@ -4,6 +4,10 @@ mkdir -p release/prev
 cp -r src/ release/prev/src/
 find release/prev/ -name "*.class" -type f -delete
 cd release
-rm 63120217-$1.zip
+
+if [ -f 63120217-$1.zip ]; then
+	rm 63120217-$1.zip
+fi
+
 zip -r 63120217-$1.zip prev/
 rm -r prev/
